refactor(cli): restructure monocore binary and improve CLI arguments
- Split monocore.rs into multiple files under bin/monocore/
- Move command handling logic into separate handlers module
- Change CLI arguments to use explicit flags instead of positional/optional args
- Make component names required arguments where applicable
- Remove ambiguous argument combinations
- Add explicit component type flags (--sandbox, --build, --group)
- Remove remote subcommand
